+++ contracts/IndexerProtein.sol
@@ -7,7 +7,6 @@ import './IndexerSeed.sol';
 //Created by Tousuke (zenodeapp - https://github.com/zenodeapp/protein-crud).
 
 contract IndexerProtein is CrudProtein {
-  // For instance an indexer group: 'Homo Sapiens', with an ID of 2 would be the second Indexer for this protein type.
   Structs.IndexerStruct public indexer;
   mapping(uint => address) seedAddresses;