+++ proteins.config.js
@@ -74,9 +74,11 @@ module.exports = {
   },
 
   queryOptions: {
-    defaultSeedSize: "3",
-    defaultLimit: "0",
+    defaultLimit: 0,
     defaultCaseSensitivity: false,
+    defaultSeedSize: 3, //Only for semi-blast
+    defaultUnion: false, //Currently only for the naive algorithm
+    defaultFormat: "protein", //Possible values: "protein", "id", "nft", "sequence" or "ipfs".
   },
 
   // Normally, inserting the same protein or seed twice would revert the tx, returning an error stating it already exists.