+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## BIG DATA ERA
 
+#### 1.2.3 (2022-08-18)
+
+- Naive algorithm now supports \*-wildcard functionality (but works as ?-wildcards, see: https://support.microsoft.com/en-us/office/examples-of-wildcard-characters-939e153f-bd30-47e4-a763-61897c87b3f4)
+- Union querying has been implemented in the naive algorithm.
+- Refactored and made the algorithms more cohesive.
+- Helper function queryProtein created and refactored the tasks in hardhat.config.js.
+- Format flag implemented, this allows us to change the return values when querying. Available values are: "protein", "nft", "id", "sequence" and "ipfs".
+- Tasks minimized to: `npx hardhat queryNaive` and `npx hardhat querySemiBlast`. Use the new --format flag to change what results you want.
+
 #### 1.2.2 (2022-08-16)
 
 - Separated Naive and Semi-Blast algorithms into two contracts.